feat: improve error handling in `STABLE_WITH_MALFORMED` mode
The `STABLE_WITH_MALFORMED` mode now has improved error handling. Instead of just a 50/50 chance to return a "valid" response or throw an exception, there is a new third branch. This third path allows to return an invalid response with non-numeric entries in the response. In addition, `NO_CONNECTION` will also throw an IOException now, instead of just having a 50% chance to do so.
- Implemented a `randomOneThirdChance` function in `RandomSimulator` to have 3 different options.
- Improved the logic in `STABLE_WITH_MALFORMED` mode to have 3 different response types.
- Improved `NO_CONNECTION` to always throw an exception.
- Improved unit tests to cover new error handling.
- `NetworkResponse.kt` now uses generic `List<Any>` to have an easier time including invalid response types.